resolve #5

+++ app/Http/Controllers/MemosController.php
@@ -11,16 +11,20 @@ class MemosController extends Controller
     // getでmemos/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
             // メモ更新順で取得
-        $memos = Memo::orderBy('updated_at', 'desc')->paginate(25);
+            $memos = $user->memos()->orderBy('updated_at', 'desc')->paginate(25);
             // メモ一覧viewでそれを表示
             return view('memos.index', [
                 'memos' => $memos,
             ]);
-        
-    }
-
+            return redirect('/');
+        }
+         return view('welcome', $data);
+    }   
     // getでmemos/createにアクセスされた場合の新規作成画面の表示処理
     public function create()
     {
@@ -50,14 +54,14 @@ class MemosController extends Controller
             'content' => 'required',
         ]);
 
-        // メッセージを作成
-        $memo = new Memo;
-        $memo->title = $request->title;
-        $memo->content = $request->content;
-        $memo->save();
+            $request->user()->memos()->create([
+            'content' => $request->content,
+            'title' => $request->title,
+        ]);
 
-        // トップページへリダイレクトさせる
+        // 前のURLへリダイレクトさせる
         return redirect('/');
+       
     }
     
     // getでmemos/id/editにアクセスされた場合の更新画面表示処理
@@ -67,11 +71,13 @@ class MemosController extends Controller
         $memo = Memo::findOrFail($id);
         
             // メモ編集をviewで表示させる
+        if (\Auth::id() === $memo->user_id) {
             return view('memos.edit', [
             'memo' => $memo,
         ]);
-        
+            }
         return redirect('/');
+        
     }
     // putまたはpatchでmemos/idにアクセスされた場合の更新処理
     public function update(Request $request, $id)
@@ -87,14 +93,17 @@ class MemosController extends Controller
         
         
         // メモを更新する
-        $memo->title = $request->title;
-        $memo->content = $request->content;
-        $memo->save();
+        
+        if (\Auth::id() === $memo->user_id) {
+            $memo->title = $request->title;
+            $memo->content = $request->content;
+            $memo->save();
             
+        }
         
         
-        // トップページへリダイレクトさせる
         return redirect('/');
+            
         
     }
 
@@ -104,8 +113,11 @@ class MemosController extends Controller
         
         // idの値でメモを検索して取得する
         $memo = Memo::findOrFail($id);
-            // メモを削除
+          
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $memo->user_id) {
             $memo->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
